extract MediaPreserter
