+++ Picroup-iOS/Config.swift
@@ -10,6 +10,6 @@ struct Config {
     static let baseURL = developBaseURL
     static let productionBaseURL = "https://api.picroup.com:4000"
     static let developBaseURL = "https://home.picroup.com:3500"
-    static let maxDiskImageCacheSize: UInt = 100 * 1024 * 1024
-    static let maxDiskVideoCacheSize: UInt = 150 * 1024 * 1024
+    static let maxDiskImageCacheSize: UInt = 100 * 1024 * 1024 // 100 MB
+    static let maxDiskVideoCacheSize: UInt = 150 * 1024 * 1024 // 100 MB
 }
