+++ Picroup-iOS/UseCases/UIKit/Cell/ImageDetailCell/ImageDetailCell.xib
