+++ Picroup-iOS/UseCases/UIKit/Cell/RankMediumCell/RankMediumCell.swift
